buildcontrol: move actions and reducers into their own package
