+++ internal/store/buildcontrols/actions.go
@@ -1,4 +1,4 @@
-package buildcontrol
+package buildcontrols
 
 import (
 	"time"
