+++ internal/store/buildcontrols/log.go
@@ -1,4 +1,4 @@
-package buildcontrol
+package buildcontrols
 
 import (
 	"context"
